Moving definitions of reorder buffer from .c to .h file
