+++ lib/librte_reorder/rte_reorder.c
@@ -2,8 +2,6 @@
  * Copyright(c) 2010-2014 Intel Corporation
  */
 
-#include <inttypes.h>
-#include <string.h>
 
 #include <rte_string_fns.h>
 #include <rte_log.h>
@@ -22,31 +20,9 @@ static struct rte_tailq_elem rte_reorder_tailq = {
 };
 EAL_REGISTER_TAILQ(rte_reorder_tailq)
 
-#define NO_FLAGS 0
-#define RTE_REORDER_PREFIX "RO_"
-#define RTE_REORDER_NAMESIZE 32
-
 /* Macros for printing using RTE_LOG */
 #define RTE_LOGTYPE_REORDER	RTE_LOGTYPE_USER1
 
-/* A generic circular buffer */
-struct cir_buffer {
-	unsigned int size;   /**< Number of entries that can be stored */
-	unsigned int mask;   /**< [buffer_size - 1]: used for wrap-around */
-	unsigned int head;   /**< insertion point in buffer */
-	unsigned int tail;   /**< extraction point in buffer */
-	struct rte_mbuf **entries;
-} __rte_cache_aligned;
-
-/* The reorder buffer data structure itself */
-struct rte_reorder_buffer {
-	char name[RTE_REORDER_NAMESIZE];
-	uint32_t min_seqn;  /**< Lowest seq. number that can be in the buffer */
-	unsigned int memsize; /**< memory area size of reorder buffer */
-	struct cir_buffer ready_buf; /**< temp buffer for dequeued entries */
-	struct cir_buffer order_buf; /**< buffer used to reorder entries */
-	int is_initialized;
-} __rte_cache_aligned;
 
 static void
 rte_reorder_free_mbufs(struct rte_reorder_buffer *b);
